refactor: refactor app hooks
