+++ src/utils/const.ts
@@ -1,6 +1,7 @@
 export const BLOCK_POLLING_TIME = 4000
 export const MAX_CONFIRMATION = 1000
 export const BLOCKCHAIN_ALERT_POLLING_TIME = 10000
+export const RESIZE_LATENCY = 500
 
 export const CachedKeys = {
   // Block List
